♻️ refactor: ProfilePage 내 페이지 / 상대방 페이지 구분
